requirements.txt
Flask==2.3.3
pandas==2.0.3
scikit-surprise==1.1.1
